blank field ignore case

+++ EmpApp.py
@@ -517,39 +517,41 @@ def updateEmployee():
     pri_skill = request.form['pri_skill']
     location = request.form['location']
 
-    update_sql = "UPDATE employee SET first_name = %s, last_name = %s, pri_skill = %s, location = %s WHERE emp_id = %s"
-    cursor = db_conn.cursor()
+    if (first_name != "" && last_name != "" && pri_skill != "" && location != "") {
+        update_sql = "UPDATE employee SET first_name = %s, last_name = %s, pri_skill = %s, location = %s WHERE emp_id = %s"
+        cursor = db_conn.cursor()
 
-    cursor.execute(update_sql, (first_name, last_name, pri_skill, location, emp_id))
-    db_conn.commit()
-    cursor.close()
-
+        cursor.execute(update_sql, (first_name, last_name, pri_skill, location, emp_id))
+        db_conn.commit()
+        cursor.close()
+    }
     return render_template("ManageEmp.html")
 
 @app.route("/deleteEmp", methods=['POST'])
 def deleteEmployee():
     emp_id = request.form['emp_id']
 
-    delete_sql_1 = "DELETE FROM employee WHERE emp_id=%s"
-    delete_sql_2 = "DELETE FROM attendance WHERE emp_id=%s"
-    delete_sql_3 = "DELETE FROM payroll WHERE emp_id=%s"
-    delete_sql_4 = "DELETE FROM performance WHERE emp_id=%s"
-    #delete_sql = "DELETE FROM employee LEFT JOIN attendance ON employee.emp_id = attendance.emp_id LEFT JOIN payroll ON employee.emp_id = payroll.emp_id LEFT JOIN performance ON employee.emp_id = performance.emp_id WHERE employee.emp_id=%s"
-    cursor1 = db_conn.cursor()
-    cursor2 = db_conn.cursor()
-    cursor3 = db_conn.cursor()
-    cursor4 = db_conn.cursor()
+    if (emp_id != "") {
+        delete_sql_1 = "DELETE FROM employee WHERE emp_id=%s"
+        delete_sql_2 = "DELETE FROM attendance WHERE emp_id=%s"
+        delete_sql_3 = "DELETE FROM payroll WHERE emp_id=%s"
+        delete_sql_4 = "DELETE FROM performance WHERE emp_id=%s"
+        cursor1 = db_conn.cursor()
+        cursor2 = db_conn.cursor()
+        cursor3 = db_conn.cursor()
+        cursor4 = db_conn.cursor()
 
-    cursor1.execute(delete_sql_1, (emp_id))
-    cursor2.execute(delete_sql_2, (emp_id))
-    cursor3.execute(delete_sql_3, (emp_id))
-    cursor4.execute(delete_sql_4, (emp_id))
-    db_conn.commit()
-    cursor1.close()
-    cursor2.close()
-    cursor3.close()
-    cursor4.close()
+        cursor1.execute(delete_sql_1, (emp_id))
+        cursor2.execute(delete_sql_2, (emp_id))
+        cursor3.execute(delete_sql_3, (emp_id))
+        cursor4.execute(delete_sql_4, (emp_id))
 
+        db_conn.commit()
+        cursor1.close()
+        cursor2.close()
+        cursor3.close()
+        cursor4.close()
+    }
     return render_template("ManageEmp.html")
 
 if __name__ == '__main__':
